[Refactor] Move fla to third party (#48500)
