+++ tests/kernels/core/test_fused_rms_norm_gated.py
@@ -7,7 +7,7 @@ matching the eager triton kernel output."""
 import pytest
 import torch
 
-from vllm.model_executor.layers.fla.ops.kda import FusedRMSNormGated
+from vllm.third_party.flash_linear_attention.ops.kda import FusedRMSNormGated
 from vllm.utils.torch_utils import set_random_seed
 
 DTYPES = [torch.bfloat16]
