+++ tests/kernels/test_fused_gdn_post_conv.py
@@ -10,7 +10,7 @@ import pytest
 import torch
 import torch.nn.functional as F
 
-from vllm.model_executor.layers.fla.ops.fused_gdn_prefill_post_conv import (
+from vllm.third_party.flash_linear_attention.ops.fused_gdn_prefill_post_conv import (
     fused_post_conv_prep,
 )
